+++ internal/app/repo/method_add.go
@@ -1,0 +1,10 @@
+package repo
+
+import (
+	"context"
+	"github.com/ozonmp/srv-verification-api/internal/model"
+)
+
+func (r repo) Add(ctx context.Context, event []model.VerificationEvent) error {
+	return nil
+}
